requirements.txt
flask
scikit-learn
mysql-connector-python
pandas